refactor(frontend): extract useVersionCompare composable (W6)
Move the version-compare interaction (compare state + selectedHistoryVersionIds/
canCompareSelected computeds + handleVersionSelectionChange/loadVersionCompare/
compareSelectedVersions/stepVersionCompare/clearVersionHistorySelection) into
composables/useVersionCompare.js, smoke-gated. workflow/versionList/changeMode
injected; selectedHistoryVersionIds returned for isVersionSelectable. Version
mutations (rollback/delete) and publish-record dialog stay in the component.
Also drops the dead resolveLeftVersionId (0 callers).

Verified: WorkflowDetail mount smoke + full suite (27 files / 116 passed);
npm run build, npm run lint (0 errors). WorkflowDetail.vue 2240 -> 2131.
